Improve initial waiting for node information with informers.
Before, we were doing periodic calls to k8s control plane with
exponential backoff. Now, we use informers to get updates about
Node/CiliumNode, which allows us to react much faster after
cilium-operator assigns CIDR to node.

Fixes: #26833
